Committing changes for user registration and spring application.
