--- src/test/java/com/sapient/SpringAppApplicationTests.java
@@ -1,19 +1,0 @@
-package com.sapient;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.shadowpact.SpringAppApplication;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = SpringAppApplication.class)
-@WebAppConfiguration
-public class SpringAppApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
